backend/pyproject.toml
[project]
name = "papertales-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.135.1",
    "google-adk>=1.26.0",
    "google-cloud-firestore>=2.23.0",
    "google-cloud-storage>=3.9.0",
    "google-cloud-texttospeech>=2.34.0",
    "httpx>=0.28.1",
    "pdfplumber>=0.11.9",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pymupdf>=1.27.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.41.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.4",
]
